requirements.txt
Flask-SQLAlchemy==1.0
